remove database types and be consistent with api types imports
