+++ util/util.go
@@ -4,9 +4,11 @@ package util
 
 import (
 	"html"
+	"net/url"
 	"strings"
 
 	"github.com/go-vela/types/library"
+	"github.com/microcosm-cc/bluemonday"
 
 	"github.com/gin-gonic/gin"
 	"github.com/go-vela/types"
@@ -127,4 +129,35 @@ func CheckAllowlist(r *library.Repo, allowlist []string) bool {
 	}
 
 	return false
+}
+
+// Sanitize is a helper function to verify the provided input
+// field does not contain HTML content. If the input field
+// does contain HTML, then the function will sanitize and
+// potentially remove the HTML if deemed malicious.
+func Sanitize(field string) string {
+	// create new HTML input microcosm-cc/bluemonday policy
+	p := bluemonday.StrictPolicy()
+
+	// create a URL query unescaped string from the field
+	queryUnescaped, err := url.QueryUnescape(field)
+	if err != nil {
+		// overwrite URL query unescaped string with field
+		queryUnescaped = field
+	}
+
+	// create an HTML escaped string from the field
+	htmlEscaped := html.EscapeString(queryUnescaped)
+
+	// create a microcosm-cc/bluemonday escaped string from the field
+	bluemondayEscaped := p.Sanitize(queryUnescaped)
+
+	// check if the field contains html
+	if !strings.EqualFold(htmlEscaped, bluemondayEscaped) {
+		// create new HTML input microcosm-cc/bluemonday policy
+		return bluemondayEscaped
+	}
+
+	// return the unmodified field
+	return field
 }
